requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Brotli==1.1.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.2.0
Django==5.1.2
django-filter==24.3
django-ipython==0.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.1.0
gunicorn==23.0.0
inflection==0.5.1
ipdb==0.13.13
ipython==8.28.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
wcwidth==0.2.13
whitenoise==6.7.0
